Avoid TSAN issue
Motivation:
`swift-crypto` currently fails to compile with TSAN enabled in 5.10 and 6.0 due to a TSAN error stemming from the use of `consuming` in the `finalize()` method of `_CryptoExtras/AES/CMAC`.

Modifications:
Like #384, the `consuming finalize()` method in `_CryptoExtras/AES/CMAC` is now wrapped inside a `#if compiler(>=6.1)` condition. A non-consuming variant is used otherwise.

Result:
`swift-crypto` can successfully compile with TSAN enabled in 5.10 and 6.0.